+++ CHANGELOG.md
@@ -9,5 +9,9 @@
 - Optimize Obfuscator class
 
 ## 1.0.3
-- Trim password to 32 char length if key ore than 32 char
+- Auto trim password to 32 char length if key ore than 32 char
 - Add note single quote usage in README.md
+
+## 1.0.4
+- Add help flag options
+- Add option AESmode for encryption commands